add update balance mapping from request to entity

+++ features/reservations/entity.go
@@ -16,8 +16,8 @@ type ReservationEntity struct {
 	RoomID       uint
 	UserID       uint
 	RoomName     string
-	Price        float64
-	Balance      float64
+	Price        int
+	Balance      int
 	CreatedAt    time.Time
 	UpdatedAt    time.Time
 }
@@ -37,13 +37,13 @@ type ReservationRequest struct {
 }
 
 type ReservationResponse struct {
-	ID           uint    `json:"id"`
-	RoomName     string  `json:"room_name"`
-	CheckInDate  string  `json:"check_in"`
-	CheckOutDate string  `json:"check_out"`
-	Price        float64 `json:"price"`
-	TotalNight   int     `json:"total_night"`
-	TotalPrice   int     `json:"total_price"`
+	ID           uint   `json:"id"`
+	RoomName     string `json:"room_name"`
+	CheckInDate  string `json:"check_in"`
+	CheckOutDate string `json:"check_out"`
+	Price        int    `json:"price"`
+	TotalNight   int    `json:"total_night"`
+	TotalPrice   int    `json:"total_price"`
 }
 
 //go:generate mockery --name ReservationService_ --output ../../mocks
