+++ features/reservations/models/reservation.go
@@ -15,6 +15,6 @@ type Reservation struct {
 	RoomID       uint
 	UserID       uint
 	RoomName     string
-	Price        float64
-	Balance      float64
+	Price        int
+	Balance      int
 }
